Using your experience with building an inventory, turn-based combat and building a game of pool construct and RPG system where the player’s avatar needs to fight enemies in a turn-based combat scenario. They will then gain items from those scenarios, which they can sell at the shop. This allows the player to buy better items or gear (2-3 pieces) from the shop, which increases their statistics (stats) to fight more difficult enemies.
Below is a detailed but not necessarily exhaustive list of the requirements of the project.

The avatar is required to have modifiable states based on the equipment they have.
The avatar is required to have base stats, which are used in combat when the avatar has no gear on.
The stats are the same as your previous assignment: attack strength, heal ability, defence, and health.
The enemies must increase in difficulty and their stats must be obviously displayed to indicate the difficulty.
The details in combat must be clearly and coherently displayed.
Begin the game in the combat scene, and then when the player either loses or wins, you must send the player to the shop to buy gear (provide the avatars with money at the beginning).
When the player buys gear you it must be immediately applied to the player.
If you wish to store items the player has already bought, but will not be using, you can optionally store the items for later use. Obviously, you will require a feature for the player to swap out different items in this case.
You are welcome to construct the shop interaction as you wish, as long as the process and feature is intuitive and responsive.
The shop must provide an option to return to combat (i.e. at least two combat encounters for demonstration purposes).