Develop and design turn based combat experience where the player's avatar either fights 2-3 enemies in one combat session, or fights 2-3 enemies in a row. The player avatar and the enemies are expected to have several options including attacking enemies, blocking attacks, or healing. The game is expected to have an elegant opening to the battle, visuals communicating the different actions and the response from the system. These indications include when the player is allowed to interact with the system, when the enemy is attacking and when the player has either lost or won the game.
Place the player avatar and enemy on opposite sides of the screen.
Add a menu for the player to choose an action when it is their turn to execute an action, these actions will include attacking, blocking and healing.
You can optionally add a menu item to allow the player to use multiple items, or multiple abilities, you will need to display a new UI modal to show the possible options for the player.
Enemies must automatically execute their attacks but you must ensure that the process is visually clear, either through visual effects or UI. Consider displaying the system completely through text - via the console, or a UI text box - before demonstrating it in other ways.
Ensure the player is aware when the combat is over using UI.
You are welcome to optionally add audio effects.
The player avatar and the different enemies should have different statistics (stats) associated with them.
Some stats to take into consideration:
Health
Attack Strength
Defence Amount
Healing Ability
For your own clarity, use a spreadsheet to decide which data to apply to the stats. Your prototype is not expected to perfectly balanced or exceptionally well designed, but it is expected to demonstrate the data moving through your system.
These stats are used to illustrate how powerful an attack, block, or healing ability is.
The player should know the health of the avatar and the enemies.
You can optionally provide further insight to the player regarding the stats of the enemies.