An effective user interface is the basis for most applications, especially video games. This is what allows the game to transfer information to the player, and provide the player with a means of interacting with the system. You are required to build the user interface for an inventory system that consists of   a shop to buy items  a bag to hold bought items and  a chest for storing items from the bag when the bag is full and can no longer contain any items.
There must be three views: a shop, a bag, and a chest.
The player must be able to open, at the same time either the bag and the shop or the bag and the chest.
They open up side-to-side as two sets of inventory systems, you are allowed to be creative in the design of your UI, but be sure to playtest and get feedback from tutors and peers
When the bag and the shop are open, the player can select different items and buy different amounts of the item.
This requires there to be an option in the UI for adjusting the amount of an individual item being bought. For example, via text input, sliders, decreasing or increasing buttons etc.
There should be a limit to the number of items that can be stored in the bag.
Ensure there is a finite amount of money. This also means that the shop items must have prices when purchasing,  and you must prevent the player from over spending.
When the bag and the chest are open, the player can drag and drop items from the bag to the chest for storage. You are welcome to be creative with this interaction, but ensure that it is clear.