Create 2375681_Anele_MzDesign, develop and build a single pool (billiards) game that has a complete game loop.      The game is usually played on a rectangle, with cue sticks, a single cue ball and a number of other balls. The table is boarded by pockets, which players will try to put the other balls in with the cue stick and cue ball.   Generally, in a game of pool a player will start by hitting a cue ball with a cue stick from the foot of the table towards a collection of balls at the head of the table. You will only be expected to implement a single player version of the game, but the balls are often divided into two sets, as the games are usually played by two players. There is also sometimes a final ball (the black 8 ball in pool, for example). The players proceed to use the cue stick and cue ball to in attempt to sink the other balls. There are various rules regarding players’ turns, penalties and points.     You are required to build a game system that provides the physics, input, game loop system and minimal user interface to allow for the game to be played. You are allowed to be creative with some of the mechanics of the game, but you will be primarily assessed on your ability to implement the physics response of the prototype and you are also expected to provide supporting systems for expressing that physics implementation effectively. Lastly, please ensure that your game effectively communicates the game state where necessary.aza_Physics Assignment-1.zip
Below are some suggestions and insights into how you can approach the building of your project.

Use the mouse to control the speed and direction of the cue ball. Click on and hold the mouse to select the cue ball, and then drag away from it to indicate the opposite direction (remember in Vectors, the vector from A to B, known as AB is B minus A). (Hint: IPointerDragHandlerLinks to an external site.)
When moving the mouse you can also check the distance from the cue ball. You can use this value to control how much force is applied to the cue ball when you release the mouse.
When you release the mouse, and fire off the cue ball, prevent input from the player.
Provide some of the rigidbodies with some bounciness (using a Physics Material) on them so they can impart some of their energy to the other balls they hit, but some friction or drag so they can slow down after being hit.
Add the physics material to the balls and to the cushions of the pool table.
Ensure all your objects are similar sizes to the real-world counterparts, to ensure the physics responds correctly. Use the default cube in Unity to get a sense of how big 1 unity metre is, as the cube is 1m2 in volume.
Keep track of the velocity of all the balls. When the total velocity of all the balls is zero, or very close to zero, allow input from the player again. Use a list and loop through them, when they should be moving.
Add trigger colliders to the pocket corners of the rectangle and in the middle of the longest part of the rectangle. When the balls collide with the triggers, destroy the balls, and provide the player a reward  (either give the players points, or another turn and so on), or a penalty if they weren’t supposed to sink that ball.
Use UI to provide the player with insight into the game state. Such as whether they have  successfully sunk a shot, or have a penalty and any other information.
Use a Raycast to indicate where the cue ball is going to land, to provide aim guidance for the player. Remember, the RaycastHit property tells you what point on the collider the raycast hit.