3 X Enemy Scamp Sheets: Krita File Format (.kra)  These should include: Silhouettes, volumes/shapes, a brief  (1-2 sentence) character description                                3 X Enemy Turnarounds: Krita File Format (.kra)  1 Turnaround (see below) for each of your 3 enemy designs
Enemy Designs:

You are required to design 3 enemies:

A walking enemy
A jumping enemy
A shooting enemy
Each enemy’s visual design should communicate its function in the game. The player should be able to read from your design what the basic mechanics that govern the enemy’s behaviour will be. When you are designing your enemy think about how it will appear on screen, consider its visual legibility (size, relative scale, colour, overall linear motif, etc.) as well as any functional concerns that will inform its design. For example: if your shooting enemy can target the player does it need to be able to aim at the player? If so how will this affect the design? Does it need a flexible arm, or a pivot for its weapon to rotate on? Will this mean that the weapon needs to be a separate GameObject?

For each enemy you are required to produce:

A scamp sheet
These should consist of a minimum of 10 variations of the enemy in silhouette. They should focus on the volumes and basic shapes of the character, not the details. Remember to consider the implementation of the character in your design, enemies must work as elements in your game not just look pretty. When designing enemies, focus on the overall read of the character and make sure that each of your designs is distinct from the others.
An enemy turnaround
You will pick ONE of your scamps for each of the 3 enemy variations to be your final design. This design should be cleaned up and redrawn from 3 different angles:
Front
Side
Back
The turnaround should be drawn to a higher level of completion than the scamps and should show the details of the enemy’s design as well as well as some initial colour choices for the enemy. The colours are not final at this stage, they may change later.