Character Rigging and Animation
Rigging:

You will be required to rig your main character sprite using the 2D bone animation system in Unity. Your character should be rigged for animation. Remember that although this assignment only requires you to animate a walk cycle for your character, your rig should support all the animations that you have planned for your character (like idle, jump, etc.). Your character should be cut-out and assembled in their default idle pose. Remember to edit the sorting layers in Krita so that they are in the right order, and to export as a .PSD and the manually convert the file to .PSB format.

Once this is done, you will need to import the file and use Unity’s skinning editor to build your rig. Remember to name the bones in your hierarchy appropriately (failure to do so will result in a penalty to your marks). Once this is done you will need to generate and weight spritemeshes for the rig. Be careful to weight your spritemeshes appropriately and check the deformations of your joints to make sure that the deformation is smooth and usable.

Animation:

Once your character is rigged you are required to produce a walk-cycle animation for the character. This is done by manipulating your rig whilst recording an animation clip in Unity’s Animation tab. Your animation should cycle, i.e. the first and last frames of the animation should be the same, so that when the animation is played it appears to play continuously without interruption. Your animation should be presented through Unity’s default scene camera so that when the play button is pressed the animation runs in the Game window.