Deliverables:        1 X Blender scene file (.blend format) containing your Low-Poly asset with UV coordinates and a normal map applied to it and a colour map.                                  1 X Normal Map in .png format                                  1 X Albedo Map in .png format
For this assignment you will be required to generate UV coordinates, bake a normal map and paint a texture (albedo) map for the asset that you built in the previous assignment. You will use the UV tools in Blender, as well as its normal map baking and texture painting system (to be covered in tutorials in class) to produce the aforementioned modifications.

UVs, normals and textures are an essential part of the preparation of any 3D asset as they allow high frequency detail on the asset to be represented as texture data that can be processed efficiently at runtime. These maps are required for colour and lighting information to be correctly assigned to the surface of the mesh when it is rendered.

For this assignment you will be required to:

Generate UV coordinates for your 3D asset.
Use Blender’s baking tools to generate a Normal Map for your asset
Paint a colour (albedo) map for your asset
Remember:

The focus of this exercise is on your technical ability to create game assets, and not artistic skill. As such the criteria you will be marked on will be the following:

Good UV mapping and usage of the space on your UV tile.
Your UVs should be arranged to make maximum use of the space on the tile, and should not exhibit excessive stretching or compression.
Remember to consider UV seam placement and texel density when laying out your UVs
Good Normal Map baking
No errors in the normal map (look out for yellow or black areas in your bake)
Good mapping to the object (this is based on the quality of your UVs)
Good Albedo Map:
Appropriate texturing of the asset (so no just throwing some colour onto your asset and calling it done)
Making sure that the texture map lines up with the details on your normal map