Rigging
You are required rig a character asset for this assignment. You will prepare the asset for animation in Assignment 4, using the rigging tools that have been covered in class. You are required to produce a fully functioning rig, including functioning controls for animators to use. The character should also be weight painted so that it deforms appropriately. For this assignment you may use the Rigify plugin’s skeleton as a starting point for your rig (like we did in class), however you may NOT use the plugin to generate your final rig.

You may use either the bLAMBO or Chichi characters provided in the Assets folder on Ulwazi, or your own character that you created in previous assignments for this course.

For this assignment you will be required to:

Prepare the character asset for rigging in Blender.
Create a deformation hierarchy for the asset, including an IK/FK switching system for the legs (the arms can be FK only).
Weight-paint the character appropriately so that it exhibits good deformation when animated.
Create any necessary control bones for the IK, etc.
Create animator controls to use when animating the asset.