Walk Cycle
You are required to animate the character asset provided for this assignment (either bLAMBO or RJ). You will prepare the asset for animation in Blender using the tools that have been covered in class. You will then animate a walk cycle for the asset. The walk should be about 32 frames long. You should save your animation as separate Actions so that they can be easily selected and played in the NLA (Non-linear Animation) Editor in Blender (this will be covered in class). You should save and submit your work as a .blend scene file with the animation saved in the scene.

For this assignment you will be required to:

Prepare your Blender scene for animation, including:
Setting up your layout to have the Graph Editor, Dope Sheet and Timeline
Saving out your individual animation as an Action for the NLA Editor.
Animate a +/- 32 frame walk cycle for the character asset.