Platformer Game Jam
You are tasked with designing and developing a platformer game in the Thursday 2-5 slot. Note that this is a micro-game jam, so we expect you to show focus. We are looking for a 10 - 15 second play experience demonstrating a clear relationship between the game's actions and challenges.  This relationship should be pushed and explored. Take the time to try to experiment with your understanding of platformers. You may take inspiration from as many games as you'd like but need to demonstrate an understanding of why you choose them.