Assignment 1 - Prototype 1
You are to create a turn-based role-playing game (RPG) for your first assignment. The game/prototype can include and use other genres, but at its core, the main systems and mechanics must be that of a Turn Based Role Playing Game.

In addition to a playable build, you are to submit a Game Design Document comprised of:

A Genre and subgenre analysis:

Doing research define the genre and subgenres of the game you are creating through its conventions, characteristics and tropes. Select 2 specific games within the genre and identify core systems and mechanisms which ground these games as part of the chosen genre. You will be using these as inspiration for the prototype/game. You may use more than two games but it is suggested for major mechanics and systems to base them on fewer examples so that you have a sharper focus. Place emphasis on defining and developing a vocabulary which best informs the function and relationship of the genre and chosen systems.

Hypothesis / Interrogation / Design goal:

From your understanding of the genres outline a hypothesis/interrogation/design goal that tests your understanding of the genres in creating a designed experience. These can take form of a variety of focuses from clarifying understanding of the genre and systems (developing craft) to exploring alt and fringe takes on the genre
(subverting or experimenting).

Design notes & process:

Include design flows, system structures, value chains, ideal screen shots, level designs etc. to convey as much of your thoughts as possible.  Also inlcude your design process which focuses on the development processes, decision making and changes you made. These are to be justified and your thought explained.

Reflection:

Reflect hypothesis/interrogation by using your process and design to address your findings.

You are also to submit your game design spreadsheets:

These are comprised of your game systems, data sets, and other tools prototyped using spreadsheets.