Semester Project
For your exam, select one of your three previous builds and iterate on it. Use your previous project as a foundation to further build on your idea and take it further. Your project should illustrate and clear reflection of your previous project, how you wish to take the project forward and a comparative reflection of b both the exam assignment and the initial assignment.  Use your design document, sheets and builds to illustrate this. Both builds will be viewed when marked.

Please do not hesitate to book a consultation with me or the teaching complement.

Submission requirements:

The game should make use of difficulty progression, a variety of enemies and levels to illustrate you design intentions. This requires a clear and concise focus that you want to explore and unpack.

A Genre and subgenre analysis:

Doing research define the genre and subgenres of the game you are creating through its conventions, characteristics and tropes. Select 2 specific games within the genre and identify core systems and mechanisms which ground these games as part of the chosen genre. You will be using these as inspiration for the prototype/game. You may use more than two games but it is suggested for major mechanics and systems to base them on fewer examples so that you have a sharper focus. Place emphasis on defining and developing a vocabulary which best informs the function and relationship of the genre and chosen systems.

Hypothesis / Interrogation / Design goal:

From your understanding of the genres outline a hypothesis/interrogation/design goal that tests your understanding of the genres in creating a designed experience. These can take form of a variety of focuses from clarifying understanding of the genre and systems (developing craft) to exploring alt and fringe takes on the genre
(subverting or experimenting).

Design notes & process:

Include design flows, system structures, value chains, ideal screen shots, level designs etc. to convey as much of your thoughts as possible.  Also inlcude your design process which focuses on the development processes, decision making and changes you made. These are to be justified and your thought explained.

Reflection:

Reflect hypothesis/interrogation by using your process and design to address your findings.

You are also to submit your game design spreadsheets:

These are comprised of your game systems, data sets, and other tools prototyped using spreadsheets.

Submit builds for the initial assignment and exam assignment.

Assignment Build

Exam Build